+++ server/database/database.service.js
@@ -5,7 +5,7 @@
 'use strict'
 
 /** Common functions for CRUD operations */
-module.exports = {
+module.exports = Object.freeze({
   /**
    * Create a query for find function of mongoose.
    * @param {Object} model - mongoose model.
@@ -36,5 +36,35 @@ module.exports = {
 
     // Return the query
     return cursor.exec()
+  },
+
+  /**
+   * Create a query for findOne function of mongoose.
+   *
+   * @param {Object} model - mongoose model.
+   * @param {Object} query - mongodb selector.
+   * @param {Object} projection - optional fields to return of each doc.
+   * @param {Array/Object} populate - An array or object of mongoose populate
+   *  object:
+      {
+        path: 'fans',
+        match: { age: { $gte: 21 }},
+        select: { name: 1 , _id: 0}, // or 'name -_id'
+        options: { limit: 5 }
+      }
+   *
+   * @returns {Promise} mongoose promise
+   */
+  findingOne(model, query, projection = {}, populate = []) {
+    const cursor = model.findOne(query, projection)
+    populate = Array.isArray(populate) ? populate : [populate]
+
+    // Check if populate
+    for (let pathObj of populate) {
+      cursor.populate(pathObj)
+    }
+
+    // Return the query
+    return cursor.exec()
   }
-}
+})