+++ server/database/index.js
@@ -4,10 +4,10 @@
 const dbPool = require('./database.pool')
 const dbService = require('./database.service')
 
-module.exports = {
+module.exports = Object.freeze({
   /** Pool of connections to application databases */
   dbPool,
 
   /** Common actions for crud operations */
   dbService
-}
+})