+++ server/response-error/index.js
@@ -1,0 +1,4 @@
+/**
+ * @module response-error
+ */
+module.exports = require('./response-error.service')